Refactor OpenBSD-SoftRAID format
